+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Fix CSV import by adding a newline to the INSERT statement plausible/analytics#4172
+
 ## v2.1.0 - 2024-05-23
 
 ### Added